Extract log subscriber for locking extensions
