+++ lib/active_record/locking_extensions.rb
@@ -1,5 +1,4 @@
 # encoding: utf-8
-require "active_support/log_subscriber"
 require "active_support/notifications"
 
 module ActiveRecord
@@ -81,27 +80,6 @@ module ActiveRecord
         end
       end
     end
-
-    class LogSubscriber < ActiveSupport::LogSubscriber
-      def deadlock_restart(event)
-        info "Deadlock causing restart"
-        debug event[:exception]
-      end
-
-      def deadlock_retry(event)
-        info "Deadlock causing retry"
-        debug event[:exception]
-      end
-
-      def duplicate_ignore(event)
-        info "Duplicate ignored"
-        debug event[:exception]
-      end
-
-      def logger
-        ActiveRecord::Base.logger
-      end
-    end
   end
 
   # Raise this inside a restartable_transaction to retry the transaction from the beginning.
@@ -112,4 +90,3 @@ end
 
 
 ActiveRecord::Base.extend(ActiveRecord::LockingExtensions)
-ActiveRecord::LockingExtensions::LogSubscriber.attach_to :active_record
